requirements.txt
# Core Deep Learning (Python 3.13 Compatible - Verified Dec 2025)
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
ultralytics>=8.3.0  # YOLOv8 - v8.3.240 supports Python 3.13
opencv-python>=4.8.0
pillow>=10.0.0
grad-cam  # Correct package name (not pytorch-grad-cam)

# NLP & Transformers
transformers>=4.30.0  # v4.57.3 supports Python 3.13
tokenizers>=0.15.0
sentencepiece>=0.1.99

# Explainability
shap>=0.50.0  # v0.50.0 supports Python 3.13
lime>=0.2.0.1

# Metadata & EXIF
PyExifTool>=0.5.6

# API Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
requests>=2.31.0
tqdm>=4.66.0

# Logging & Monitoring
structlog>=24.0.0
python-json-logger>=2.0.0
prometheus-client>=0.19.0

# Image Processing
albumentations>=1.3.0
